Initial addition to log deduplication experiment
This commit shouldn't be used except in experimental
settings or until it is supplanted by more fixes to correct it.

This adds log deduplication based on a near-duplicate
analysis of formatted messages (without node or timestamp)
over a given period of time. The objective is to allow
long chains of repeated events in a short burst of time
to be reduced to smaller quantities of text transmitted
and buffered to the error logger. This should allow more
resilience to overload situations, and allow to take less
place during long cascades of failures.

TODO:
- have more tests
- refactor code to avoid duplication between lager.erl and
  lager_dedup.erl
- figure out if the new application dependencies are a
  problem for users of lager.
- document new features.