Moving to an ETS-based implementation of deduper
This allows much better performances than a dict-based version.
Also note that most of the improvements are also available due
to moving to a 'first match' search for duplicates (below
the given treshold) rather than a 'best match' option, in order
to possibly reduce the maximum number of match attempts to be made
when logs are wildly different.

It's still not known if this is the best approach, but it did
allow to survive overload situations much better.

Also using ETS gives way to try some interesting optimizations
by making the table protected later on (i.e. do not send
an error message if 1000 are already in the table).