Improving shinglign/weights for deduping
Now able to process a lot more messages per second with as good
of a significance in simhashing.

This allows us to move away from phash2 as a hash algorithm (32
bits) to MD5 (128bits). While it copies a bit more data for each
communication, it allows much more accuracy in duplicate
detection.