Adding limits to deduplication
This allows to do an optional pre-read in ETS to find out if
the currently deduplicated log message has been seen more than
a given number of times (dictated by the 'duplicate_limit'
env variable). If it's been seen too often, the entire
negotiation between lager_deduper and its children is skipped to
avoid undesirable message sending that may take a lot of space
or processing time.

This measure is a best effort -- ETS not having locks or anything
of that kind, it is possible to get more matched messages than the
limit states.