Adding an option for quick deduplicated delivery
When logging deduplicated messages, the error handlers only
learn about the existance of an error message after they have
happened. In some cases, the time value could be set high, and
this would create latencies in fault detection.

This commit adds an option so that the first message of a group
(seen as too different to be merged in with another message) is
instantly sent to the error logger, while the following ones get
accumulated.

This will effectively change the behaviour from receiving:
  "error X (N times)"
to:
  "error X" + "error X (N-1 times)"