Fixing a race condition.
During system blockage, the gen_server:call limit of 5 seconds
could go off before giving an answer, creating an incomplete
deduplication session between lager_deduper and the error
reporter. This would in turn create dead entries that would never
get dumped in the future, effectively deleting a type of error
messages from the logs.

By setting the timeout to infinity, we allow more tolerance to
overload situations (which is when the deduper is useful) and
fix the issue entirely.