Dedup tests clean up after themselves
The duplicate_threshold value used in configuration brings a
global change behaviour in lager. When tests ran in a certain
order, deduplication tests would run before other suites that
would not reset the value given they don't need to know about
it.

This patch makes it so that deduplication tests should not
break tests in other suites.