Addresses issue #332.
For Offenders that don't have a name key, attempt to
use the id key as the source of the crashed
supervised child's name. id appears to contain the
supervisor's child id for the crashed child.

This change makes the "A supervised child has crashed!"
reports actually contain the child's name, rather
than "unknown".

We search for name first to retain compatibility with
the existing code.