Improve debug tracing support
By an unhappy accident, OTP21 added functionality for 'named' trace
functions. This has a pattern match that matches the old trace function
(a 2 tuple where the second element is itself a 2 tuple):

https://github.com/erlang/otp/pull/1781

This patch changes to using a versioned record (so we can extend the
tracing later if needed) and adds 3 more options to tracing; count,
timeout and format string.

Additionally, tracing an OTP process via start_link arguments is now
documented.