Allow formatters to display stack traces on crash
Only the last stack frame (ie. the place that cause the crash) is currently being
shown in the message provided to the formatters, if you're interested in more
information you'll need to dig into the crash.log for it.
This is not ideal when using containerized environments with custom formatters,
we want as much as data as possible to be available in the console backend.